# internal/api/server.go
package api

import (
	"context"
	"os"
	"runtime"
	"strings"
	"sync/atomic"
	"time"

	"github.com/gofiber/fiber/v2"
	"github.com/gofiber/fiber/v2/middleware/cors"
	"github.com/gofiber/fiber/v2/middleware/limiter"
	"github.com/gofiber/fiber/v2/middleware/recover"
	"github.com/javi11/altmount/internal/arrs"
	"github.com/javi11/altmount/internal/auth"
	"github.com/javi11/altmount/internal/database"
	"github.com/javi11/altmount/internal/health"
	"github.com/javi11/altmount/internal/importer"
	"github.com/javi11/altmount/internal/metadata"
	"github.com/javi11/altmount/internal/nzbfilesystem"
	"github.com/javi11/altmount/internal/nzbfilesystem/segcache"
	"github.com/javi11/altmount/internal/pool"
	"github.com/javi11/altmount/internal/progress"
	"github.com/javi11/altmount/internal/rclone"
	"github.com/javi11/altmount/internal/version"
	"github.com/javi11/altmount/pkg/rclonecli"
)

// Config represents API server configuration
type Config struct {
	Prefix string // API path prefix (default: "/api")
}

// DefaultConfig returns default API configuration
func DefaultConfig() *Config {
	return &Config{
		Prefix: "/api",
	}
}

// Server represents the API server
type Server struct {
	config              *Config
	queueRepo           *database.Repository
	healthRepo          *database.HealthRepository
	mediaRepo           *database.MediaRepository
	authService         *auth.Service
	userRepo            *database.UserRepository
	configManager       ConfigManager
	metadataReader      *metadata.MetadataReader
	metadataService     *metadata.MetadataService
	nzbFilesystem       *nzbfilesystem.NzbFilesystem
	healthWorker        *health.HealthWorker
	librarySyncWorker   *health.LibrarySyncWorker
	importerService     *importer.Service
	poolManager         pool.Manager
	arrsService         *arrs.Service
	rcloneClient        rclonecli.RcloneRcClient
	mountService        *rclone.MountService
	startTime           time.Time
	progressBroadcaster *progress.ProgressBroadcaster
	streamTracker       *StreamTracker
	fuseManager         *FuseManager
	segcacheMgr         *segcache.Manager // nil if segment cache is disabled
	ready               atomic.Bool
}

// NewServer creates a new API server that can optionally register routes on the provided mux (for backwards compatibility)
func NewServer(
	config *Config,
	queueRepo *database.Repository,
	healthRepo *database.HealthRepository,
	mediaRepo *database.MediaRepository,
	authService *auth.Service,
	userRepo *database.UserRepository,
	configManager ConfigManager,
	metadataReader *metadata.MetadataReader,
	metadataService *metadata.MetadataService,
	nzbFilesystem *nzbfilesystem.NzbFilesystem,
	poolManager pool.Manager,
	importService *importer.Service,
	arrsService *arrs.Service,
	mountService *rclone.MountService,
	progressBroadcaster *progress.ProgressBroadcaster,
	streamTracker *StreamTracker,
	segcacheMgr *segcache.Manager,
) *Server {
	if config == nil {
		config = DefaultConfig()
	}

	server := &Server{
		config:              config,
		queueRepo:           queueRepo,
		healthRepo:          healthRepo,
		mediaRepo:           mediaRepo,
		authService:         authService,
		userRepo:            userRepo,
		configManager:       configManager,
		metadataReader:      metadataReader,
		metadataService:     metadataService,
		nzbFilesystem:       nzbFilesystem,
		importerService:     importService, // Will be set later via SetImporterService
		poolManager:         poolManager,
		arrsService:         arrsService,
		mountService:        mountService,
		startTime:           time.Now(),
		progressBroadcaster: progressBroadcaster,
		streamTracker:       streamTracker,
		segcacheMgr:         segcacheMgr,
		fuseManager:         NewFuseManager(newMountFactory(nzbFilesystem, configManager, streamTracker)),
	}

	return server
}

// SetHealthWorker sets the health worker reference for the server
func (s *Server) SetHealthWorker(healthWorker *health.HealthWorker) {
	s.healthWorker = healthWorker
}

// SetLibrarySyncWorker sets the library sync worker reference for the server
func (s *Server) SetLibrarySyncWorker(librarySyncWorker *health.LibrarySyncWorker) {
	s.librarySyncWorker = librarySyncWorker
}

// SetReady sets the server as ready to accept requests
func (s *Server) SetReady(ready bool) {
	s.ready.Store(ready)
}

// IsReady returns true if the server is ready to accept requests
func (s *Server) IsReady() bool {
	return s.ready.Load()
}

// SetRcloneClient sets the rclone client reference for the server
func (s *Server) SetRcloneClient(rcloneClient rclonecli.RcloneRcClient) {
	s.rcloneClient = rcloneClient
}

// GetProgressBroadcaster returns the progress broadcaster for use by the importer service
func (s *Server) GetProgressBroadcaster() *progress.ProgressBroadcaster {
	return s.progressBroadcaster
}

// SetupFiberRoutes configures API routes directly on the Fiber app
func (s *Server) SetupRoutes(app *fiber.App) {
	app.Use("/sabnzbd", s.handleSABnzbd)

	// Stremio addon endpoints — key-based auth, no JWT required.
	// CORS must be open (*) so Stremio can install the addon from any origin.
	stremioGroup := app.Group("/stremio", cors.New(cors.Config{
		AllowOrigins: "*",
		AllowHeaders: "Origin, Content-Type, Accept",
		AllowMethods: "GET, OPTIONS",
	}))
	stremioGroup.Get("/:key/manifest.json", s.handleStremioManifest)
	stremioGroup.Get("/:key/stream/:type/:id.json", s.handleStremioAddonStream)
	stremioGroup.Get("/:key/play", s.handleStremioAddonPlay)

	api := app.Group(s.config.Prefix)

	// Public endpoints (authentication handled inside or not required)
	api.Post("/import/file", s.handleManualImportFile)
	api.Post("/arrs/webhook", s.handleArrsWebhook)
	api.Post("/nzb/streams", s.handleNzbStreams)

	cfg := s.configManager.GetConfig()

	// Apply global middleware — derive allowed CORS origins from COOKIE_DOMAIN env var,
	// explicit api.allowed_origins config, or fall back to wildcard.
	corsOrigins := "*"

	if cookieDomain := os.Getenv("COOKIE_DOMAIN"); cookieDomain != "" {
		// Allow both http and https on the configured domain
		corsOrigins = "http://" + cookieDomain + ",https://" + cookieDomain
	}

	if cfg != nil && len(cfg.API.AllowedOrigins) > 0 {
		// Explicit config overrides the env-derived value
		corsOrigins = strings.Join(cfg.API.AllowedOrigins, ",")
	}

	api.Use(cors.New(cors.Config{
		AllowOrigins:     corsOrigins,
		AllowHeaders:     "Origin, Content-Type, Accept, Authorization",
		AllowMethods:     "GET, POST, PUT, PATCH, DELETE, OPTIONS",
		AllowCredentials: true,
	}))
	api.Use(recover.New())

	// Apply JWT authentication middleware globally except for public auth routes
	// Only apply if login is required (default: true)
	loginRequired := true // Default to true if not set
	if cfg != nil && cfg.Auth.LoginRequired != nil {
		loginRequired = *cfg.Auth.LoginRequired
	}

	if loginRequired && s.authService != nil && s.userRepo != nil {
		tokenService := s.authService.TokenService()
		if tokenService != nil {
			// Define paths that should skip authentication
			skipPaths := []string{
				s.config.Prefix + "/auth/login",
				s.config.Prefix + "/auth/register",
				s.config.Prefix + "/auth/registration-status",
				s.config.Prefix + "/auth/config",
			}

			// Apply authentication middleware with skip paths
			api.Use(auth.RequireAuthWithSkip(tokenService, s.userRepo, skipPaths))
		}
	}

	// NZBDav Imports (now protected by JWT auth)
	api.Post("/import/nzbdav", s.handleImportNzbdav)
	api.Post("/import/nzbdav/reset", s.handleResetNzbdavImportStatus)
	api.Get("/import/nzbdav/status", s.handleGetNzbdavImportStatus)
	api.Delete("/import/nzbdav", s.handleCancelNzbdavImport)

	// Queue endpoints
	api.Get("/queue", s.handleListQueue)
	api.Get("/queue/stats", s.handleGetQueueStats)
	api.Get("/queue/stats/history", s.handleGetQueueHistoricalStats)
	api.Get("/queue/progress/stream", s.handleProgressStream) // SSE endpoint for real-time progress
	api.Delete("/queue/completed", s.handleClearCompletedQueue)
	api.Delete("/queue/failed", s.handleClearFailedQueue)
	api.Delete("/queue/pending", s.handleClearPendingQueue)
	api.Delete("/queue/bulk", s.handleDeleteQueueBulk)
	api.Post("/queue/bulk/restart", s.handleRestartQueueBulk)
	api.Post("/queue/bulk/cancel", s.handleCancelQueueBulk)
	api.Post("/queue/upload", s.handleUploadToQueue)
	api.Post("/queue/upload-nzblnk", s.handleUploadNZBLnk)
	api.Post("/queue/upload-by-name", s.handleSearchNZBByName)
	api.Post("/queue/test", s.handleAddTestQueueItem)
	api.Get("/queue/:id", s.handleGetQueue)
	api.Delete("/queue/:id", s.handleDeleteQueue)
	api.Post("/queue/:id/retry", s.handleRetryQueue)
	api.Post("/queue/:id/cancel", s.handleCancelQueue)
	api.Patch("/queue/:id/priority", s.handleUpdateQueueItemPriority)
	api.Get("/queue/:id/download", s.handleDownloadNZB)

	// Health endpoints
	api.Get("/health", s.handleListHealth)
	api.Post("/health/bulk/delete", s.handleDeleteHealthBulk)
	api.Post("/health/bulk/restart", s.handleRestartHealthChecksBulk)
	api.Post("/health/bulk/repair", s.handleRepairHealthBulk)
	api.Get("/health/corrupted", s.handleListCorrupted)
	api.Get("/health/stats", s.handleGetHealthStats)
	api.Delete("/health/cleanup", s.handleCleanupHealth)
	api.Post("/health/reset-all", s.handleResetAllHealthChecks)
	api.Post("/health/regenerate-symlinks", s.handleRegenerateSymlinks)
	api.Post("/health/check", s.handleAddHealthCheck)
	api.Get("/health/worker/status", s.handleGetHealthWorkerStatus)
	api.Post("/health/:id/repair", s.handleRepairHealth)
	api.Post("/health/:id/unmask", s.handleUnmaskHealth)
	api.Post("/health/:id/check-now", s.handleDirectHealthCheck)
	api.Post("/health/:id/priority", s.handleSetHealthPriority)
	api.Post("/health/:id/cancel", s.handleCancelHealthCheck)
	api.Get("/health/:id", s.handleGetHealth)
	api.Delete("/health/:id", s.handleDeleteHealth)

	// Library sync endpoints
	api.Get("/health/library-sync/status", s.handleGetLibrarySyncStatus)
	api.Get("/health/library-sync/needed", s.handleGetSyncNeeded)
	api.Post("/health/library-sync/start", s.handleStartLibrarySync)
	api.Post("/health/library-sync/cancel", s.handleCancelLibrarySync)
	api.Post("/health/library-sync/dry-run", s.handleDryRunLibrarySync)

	api.Get("/files/info", s.handleGetFileMetadata)
	api.Get("/files/active-streams", s.handleGetActiveStreams)
	api.Delete("/files/active-streams/:id", s.handleKillStream)
	api.Get("/files/streams/history", s.handleGetStreamHistory)
	api.Get("/files/export-nzb", s.handleExportMetadataToNZB)
	api.Post("/files/export-batch", s.handleBatchExportNZB)
	// Note: /files/stream is handled by StreamHandler at HTTP server level

	api.Post("/import/scan", s.handleStartManualScan)
	api.Get("/import/scan/status", s.handleGetScanStatus)
	api.Delete("/import/scan", s.handleCancelScan)
	api.Get("/import/history", s.handleGetImportHistory)
	api.Delete("/import/history", s.handleClearImportHistory)
	// System endpoints
	api.Get("/system/stats", s.handleGetSystemStats)
	api.Get("/system/health", s.handleGetSystemHealth)
	api.Get("/system/browse", s.handleSystemBrowse)
	api.Get("/system/pool/metrics", s.handleGetPoolMetrics)
	api.Post("/system/stats/reset", s.handleResetSystemStats)
	api.Post("/system/cleanup", s.handleSystemCleanup)
	api.Post("/system/restart", s.handleSystemRestart)

	// Update endpoints
	api.Get("/system/update/status", s.handleGetUpdateStatus)
	api.Post("/system/update/apply", s.handleApplyUpdate)

	api.Get("/config", s.handleGetConfig)
	api.Put("/config", s.handleUpdateConfig)
	api.Patch("/config/:section", s.handlePatchConfigSection)
	api.Post("/config/reload", s.handleReloadConfig)
	api.Post("/config/validate", s.handleValidateConfig)

	// FUSE endpoints
	api.Post("/fuse/start", s.handleStartFuseMount)
	api.Post("/fuse/stop", s.handleStopFuseMount)
	api.Post("/fuse/force-stop", s.handleForceStopFuseMount)
	api.Get("/fuse/status", s.handleGetFuseStatus)

	// Provider management endpoints
	api.Post("/providers/test", s.handleTestProvider)
	api.Post("/providers/:id/speedtest", s.handleTestProviderSpeed)
	api.Post("/providers", s.handleCreateProvider)
	api.Put("/providers/reorder", s.handleReorderProviders)
	api.Put("/providers/:id", s.handleUpdateProvider)
	api.Delete("/providers/:id", s.handleDeleteProvider)

	// Configuration-based instance endpoints
	api.Get("/arrs/instances", s.handleListArrsInstances)
	api.Get("/arrs/instances/:type/:name", s.handleGetArrsInstance)
	api.Post("/arrs/instances/test", s.handleTestArrsConnection)
	api.Get("/arrs/stats", s.handleGetArrsStats)
	api.Get("/arrs/health", s.handleGetArrsHealth)
	api.Post("/arrs/webhook/register", s.handleRegisterArrsWebhooks)
	api.Post("/arrs/download-client/register", s.handleRegisterArrsDownloadClients)
	api.Post("/arrs/download-client/test", s.handleTestArrsDownloadClients)

	// Direct authentication endpoints — rate-limited to prevent brute-force attacks
	authLimiter := limiter.New(limiter.Config{
		Max:        10,
		Expiration: 1 * time.Minute,
		KeyGenerator: func(c *fiber.Ctx) string {
			return c.IP()
		},
		LimitReached: func(c *fiber.Ctx) error {
			return c.Status(fiber.StatusTooManyRequests).JSON(fiber.Map{
				"success": false,
				"error": fiber.Map{
					"code":    "TOO_MANY_REQUESTS",
					"message": "Too many login attempts. Please wait a minute before trying again.",
				},
			})
		},
	})
	api.Post("/auth/login", authLimiter, s.handleDirectLogin)
	api.Post("/auth/register", authLimiter, s.handleRegister)
	api.Get("/auth/registration-status", s.handleCheckRegistration)
	api.Get("/auth/config", s.handleGetAuthConfig)

	// Protected API endpoints for user management (authentication already handled globally)
	api.Get("/user", s.handleAuthUser)
	api.Post("/user/refresh", s.handleAuthRefresh)
	api.Post("/user/logout", s.handleAuthLogout)
	api.Post("/user/api-key/regenerate", s.handleRegenerateAPIKey)
	api.Put("/user/password", s.handleChangeOwnPassword)
}

// Shutdown shuts down the API server and its managed resources
func (s *Server) Shutdown(ctx context.Context) {
	if s.fuseManager != nil {
		s.fuseManager.Stop()
	}
}

// handleGetActiveStreams handles GET /api/files/active-streams
//
// @Summary      Get active streams
// @Description  Returns a list of currently active file streams
// @Tags         Files
// @Produce      json
// @Param        type  query  string  false  "Filter by source type (e.g. file)"
// @Success      200  {object}  APIResponse
// @Security     BearerAuth
// @Router       /api/files/active-streams [get]
func (s *Server) handleGetActiveStreams(c *fiber.Ctx) error {
	if s.streamTracker == nil {
		return c.Status(200).JSON(fiber.Map{
			"success": true,
			"data":    []nzbfilesystem.ActiveStream{},
		})
	}

	streams := s.streamTracker.GetAll()

	// Check for filter parameter
	filterType := c.Query("type") // e.g., type=file

	if filterType == "file" {
		filteredStreams := make([]nzbfilesystem.ActiveStream, 0)
		for _, stream := range streams {
			// Assuming "API" and "WebDAV" are the desired "file being streams"
			if stream.Source == "API" || stream.Source == "WebDAV" {
				filteredStreams = append(filteredStreams, stream)
			}
		}
		streams = filteredStreams
	}

	return c.Status(200).JSON(fiber.Map{
		"success": true,
		"data":    streams,
	})
}

// getSystemInfo returns current system information
func (s *Server) getSystemInfo() SystemInfoResponse {
	uptime := time.Since(s.startTime)
	return SystemInfoResponse{
		Version:   version.Version,
		GitCommit: version.GitCommit,
		StartTime: s.startTime,
		Uptime:    uptime.String(),
		GoVersion: runtime.Version(),
	}
}

// checkSystemHealth performs a basic health check
func (s *Server) checkSystemHealth(ctx context.Context) SystemHealthResponse {
	components := make(map[string]ComponentHealth)
	overallStatus := "healthy"

	// Check database connectivity
	if _, err := s.queueRepo.GetQueueStats(ctx); err != nil {
		components["database"] = ComponentHealth{
			Status:  "unhealthy",
			Message: "Database connection failed",
			Details: err.Error(),
		}
		overallStatus = "unhealthy"
	} else {
		components["database"] = ComponentHealth{
			Status:  "healthy",
			Message: "Database connection OK",
		}
	}

	// Check health repository
	if _, err := s.healthRepo.GetHealthStats(ctx); err != nil {
		components["health_repository"] = ComponentHealth{
			Status:  "unhealthy",
			Message: "Health repository failed",
			Details: err.Error(),
		}
		if overallStatus != "unhealthy" {
			overallStatus = "degraded"
		}
	} else {
		components["health_repository"] = ComponentHealth{
			Status:  "healthy",
			Message: "Health repository OK",
		}
	}

	return SystemHealthResponse{
		Status:     overallStatus,
		Timestamp:  time.Now(),
		Components: components,
	}
}

// Library sync handler methods
func (s *Server) handleGetLibrarySyncStatus(c *fiber.Ctx) error {
	if s.librarySyncWorker == nil {
		return c.Status(fiber.StatusServiceUnavailable).JSON(fiber.Map{
			"error": "Library sync worker not available",
		})
	}

	handlers := NewLibrarySyncHandlers(s.librarySyncWorker, s.configManager)
	return handlers.handleGetLibrarySyncStatus(c)
}

func (s *Server) handleStartLibrarySync(c *fiber.Ctx) error {
	if s.librarySyncWorker == nil {
		return c.Status(fiber.StatusServiceUnavailable).JSON(fiber.Map{
			"error": "Library sync worker not available",
		})
	}

	handlers := NewLibrarySyncHandlers(s.librarySyncWorker, s.configManager)
	return handlers.handleStartLibrarySync(c)
}

func (s *Server) handleCancelLibrarySync(c *fiber.Ctx) error {
	if s.librarySyncWorker == nil {
		return c.Status(fiber.StatusServiceUnavailable).JSON(fiber.Map{
			"error": "Library sync worker not available",
		})
	}

	handlers := NewLibrarySyncHandlers(s.librarySyncWorker, s.configManager)
	return handlers.handleCancelLibrarySync(c)
}

func (s *Server) handleDryRunLibrarySync(c *fiber.Ctx) error {
	if s.librarySyncWorker == nil {
		return c.Status(fiber.StatusServiceUnavailable).JSON(fiber.Map{
			"error": "Library sync worker not available",
		})
	}

	handlers := NewLibrarySyncHandlers(s.librarySyncWorker, s.configManager)
	return handlers.handleDryRunLibrarySync(c)
}

func (s *Server) handleGetSyncNeeded(c *fiber.Ctx) error {
	handlers := NewLibrarySyncHandlers(s.librarySyncWorker, s.configManager)
	return handlers.handleGetSyncNeeded(c)
}

// handleKillStream handles DELETE /api/files/active-streams/:id
//
// @Summary      Kill an active stream
// @Description  Terminates an active file stream by its ID
// @Tags         Files
// @Produce      json
// @Param        id  path  string  true  "Stream ID"
// @Success      200  {object}  APIResponse
// @Failure      404  {object}  APIResponse
// @Security     BearerAuth
// @Router       /api/files/active-streams/{id} [delete]
func (s *Server) handleKillStream(c *fiber.Ctx) error {
	id := c.Params("id")
	if s.streamTracker == nil {
		return c.Status(404).JSON(fiber.Map{
			"success": false,
			"message": "Stream tracker not available",
		})
	}

	if s.streamTracker.KillStream(id) {
		return c.JSON(fiber.Map{
			"success": true,
			"message": "Stream termination requested",
		})
	}

	return c.Status(404).JSON(fiber.Map{
		"success": false,
		"message": "Stream not found or cannot be killed",
	})
}

// handleGetStreamHistory handles GET /api/files/streams/history
//
// @Summary      Get stream history
// @Description  Returns a history of recent file stream sessions
// @Tags         Files
// @Produce      json
// @Success      200  {object}  APIResponse
// @Security     BearerAuth
// @Router       /api/files/streams/history [get]
func (s *Server) handleGetStreamHistory(c *fiber.Ctx) error {
	if s.streamTracker == nil {
		return c.JSON(fiber.Map{
			"success": true,
			"data":    []nzbfilesystem.ActiveStream{},
		})
	}

	return c.JSON(fiber.Map{
		"success": true,
		"data":    s.streamTracker.GetHistory(),
	})
}